Dialogue Addition
Regression Dialogue Folder was added. It has a manifest that uses Content Patcher

The content.JSON tells Content Patcher to look in the Dialogue folder for changes.

When talking to Vincent or Jas for the first time, the player will get a question about whether or not they'd want ABDL lines for them. If they choose no, all ABDL changes for them will be replaced with normal dialogue. This includes reactions to the player using their pants.

Jodi has full dialogue for every day of the year.

Vincent has full dialogue for every day of the year.

Sebastion has one ABDL question for when I was learning how to do that.

Jas has the introduction question.

en.JSON was edited to have less humilation, Jodi reactions, and Vincent reactions only if the introduction question answered yes.